improved structure of the code to make development easier also continued to work on the issue #89 and #91
