+++ koalixcrm/crm/admin.py
@@ -1,720 +1,17 @@
 # -*- coding: utf-8 -*-
-from django import forms
-from django.contrib import admin
-from django.contrib import messages
-from django.contrib.admin import helpers
-from django.shortcuts import render
-from django.template.context_processors import csrf
-from django.utils.translation import ugettext as _
-from koalixcrm.plugin import *
-
-from koalixcrm.accounting.models import Account
-from koalixcrm.accounting.models import Booking
 from koalixcrm.crm.views import *
-from koalixcrm.crm.documents.calculations import *
-
-
-class ContractPostalAddress(admin.StackedInline):
-    model = PostalAddressForContract
-    extra = 1
-    classes = ['collapse']
-    fieldsets = (
-        ('Basics', {
-            'fields': (
-            'prefix', 'prename', 'name', 'addressline1', 'addressline2', 'addressline3', 'addressline4', 'zipcode',
-            'town', 'state', 'country', 'purpose'),
-        }),
-    )
-    allow_add = True
-
-
-class ContractPhoneAddress(admin.TabularInline):
-    model = PhoneAddressForContract
-    extra = 1
-    classes = ['collapse']
-    fieldsets = (
-        ('Basics', {
-            'fields': ('phone', 'purpose',)
-        }),
-    )
-    allow_add = True
-
-
-class ContractEmailAddress(admin.TabularInline):
-    model = EmailAddressForContract
-    extra = 1
-    classes = ['collapse']
-    fieldsets = (
-        ('Basics', {
-            'fields': ('email', 'purpose',)
-        }),
-    )
-    allow_add = True
-
-
-class PurchaseOrderPostalAddress(admin.StackedInline):
-    model = PostalAddressForPurchaseOrder
-    extra = 1
-    classes = ['collapse']
-    fieldsets = (
-        ('Basics', {
-            'fields': (
-            'prefix', 'prename', 'name', 'addressline1', 'addressline2', 'addressline3', 'addressline4', 'zipcode',
-            'town', 'state', 'country', 'purpose')
-        }),
-    )
-    allow_add = True
-
-
-class PurchaseOrderPhoneAddress(admin.TabularInline):
-    model = PhoneAddressForPurchaseOrder
-    extra = 1
-    classes = ['collapse']
-    fieldsets = (
-        ('Basics', {
-            'fields': ('phone', 'purpose',)
-        }),
-    )
-    allow_add = True
-
-
-class PurchaseOrderEmailAddress(admin.TabularInline):
-    model = EmailAddressForPurchaseOrder
-    extra = 1
-    classes = ['collapse']
-    fieldsets = (
-        ('Basics', {
-            'fields': ('email', 'purpose',)
-        }),
-    )
-    allow_add = True
-
-class SalesContractTextParagraph(admin.StackedInline):
-    model = TextParagraphInSalesContract
-    extra = 1
-    classes = ['collapse']
-    fieldsets = (
-        ('Basics', {
-            'fields': ('text_paragraph', 'purpose', )
-        }),
-    )
-    allow_add = True
-
-
-class SalesContractPostalAddress(admin.StackedInline):
-    model = PostalAddressForSalesContract
-    extra = 1
-    classes = ['collapse']
-    fieldsets = (
-        ('Basics', {
-            'fields': (
-            'prefix', 'prename', 'name', 'addressline1', 'addressline2', 'addressline3', 'addressline4', 'zipcode',
-            'town', 'state', 'country', 'purpose')
-        }),
-    )
-    allow_add = True
-
-
-class SalesContractPhoneAddress(admin.TabularInline):
-    model = PhoneAddressForSalesContract
-    extra = 1
-    classes = ['collapse']
-    fieldsets = (
-        ('Basics', {
-            'fields': ('phone', 'purpose',)
-        }),
-    )
-    allow_add = True
-
-
-class SalesContractEmailAddress(admin.TabularInline):
-    model = EmailAddressForSalesContract
-    extra = 1
-    classes = ['collapse']
-    fieldsets = (
-        ('Basics', {
-            'fields': ('email', 'purpose',)
-        }),
-    )
-    allow_add = True
-
-
-class SalesContractInlinePosition(admin.TabularInline):
-    model = SalesContractPosition
-    extra = 1
-    classes = ['expand']
-    fieldsets = (
-        ('', {
-            'fields': (
-            'position_number', 'quantity', 'unit', 'product', 'description', 'discount', 'overwrite_product_price',
-            'position_price_per_unit', 'sent_on')
-        }),
-    )
-    allow_add = True
-
-
-class InlineQuote(admin.TabularInline):
-    model = Quote
-    classes = ['collapse']
-    show_change_link = True
-    can_delete = True
-    extra = 1
-    readonly_fields = (
-    'description', 'contract', 'customer', 'valid_until', 'status', 'last_pricing_date', 'last_calculated_price',
-    'last_calculated_tax',)
-    fieldsets = (
-        (_('Basics'), {
-            'fields': ('description', 'contract', 'customer', 'valid_until', 'status')
-        }),
-        (_('Advanced (not editable)'), {
-            'classes': ('collapse',),
-            'fields': ('last_pricing_date', 'last_calculated_price', 'last_calculated_tax',)
-        }),
-    )
-    allow_add = False
-
-
-class InlineInvoice(admin.TabularInline):
-    model = Invoice
-    classes = ['collapse']
-    show_change_link = True
-    can_delete = True
-    extra = 1
-    readonly_fields = (
-    'last_pricing_date', 'last_calculated_price', 'last_calculated_tax', 'description', 'contract', 'customer',
-    'payable_until', 'status')
-    fieldsets = (
-        (_('Basics'), {
-            'fields': ('description', 'contract', 'customer', 'payable_until', 'status'),
-        }),
-        (_('Advanced (not editable)'), {
-            'classes': ('collapse',),
-            'fields': ('last_pricing_date', 'last_calculated_price', 'last_calculated_tax',)
-        }),
-    )
-
-    allow_add = False
-
-
-class InlinePurchaseOrder(admin.TabularInline):
-    model = PurchaseOrder
-    classes = ['collapse']
-    show_change_link = True
-    can_delete = True
-    extra = 1
-    readonly_fields = (
-    'description', 'contract', 'supplier', 'external_reference', 'status', 'last_pricing_date', 'last_calculated_price')
-    fieldsets = (
-        (_('Basics'), {
-            'fields': ('description', 'contract', 'supplier', 'external_reference', 'status')
-        }),
-        (_('Advanced (not editable)'), {
-            'classes': ('collapse',),
-            'fields': ('last_pricing_date', 'last_calculated_price')
-        }),
-    )
-    allow_add = False
-
-
-class OptionContract(admin.ModelAdmin):
-    list_display = ('id', 'description', 'default_customer', 'default_supplier', 'staff', 'default_currency')
-    list_display_links = ('id',)
-    list_filter = ('default_customer', 'default_supplier', 'staff', 'default_currency')
-    ordering = ('id', )
-    search_fields = ('id', 'contract')
-    fieldsets = (
-        (_('Basics'), {
-            'fields': ('description', 'default_customer', 'staff', 'default_supplier', 'default_currency', 'default_template_set')
-        }),
-    )
-    inlines = [ContractPostalAddress, ContractPhoneAddress, ContractEmailAddress, InlineQuote, InlineInvoice,
-               InlinePurchaseOrder]
-    pluginProcessor = PluginProcessor()
-    inlines.extend(pluginProcessor.getPluginAdditions("contractInlines"))
-
-    def create_purchase_order(self, request, queryset):
-        for obj in queryset:
-            purchase_order = obj.create_purchase_order()
-            self.message_user(request, _("PurchaseOrder created"))
-            response = HttpResponseRedirect('/admin/crm/purchaseorder/' + str(purchase_order.id))
-        return response
-
-        create_purchase_order.short_description = _("Create Purchaseorder")
-
-    def create_quote(self, request, queryset):
-        for obj in queryset:
-            quote = obj.create_quote()
-            self.message_user(request, _("Quote created"))
-            response = HttpResponseRedirect('/admin/crm/quote/' + str(quote.id))
-        return response
-
-    create_quote.short_description = _("Create Quote")
-
-    def create_invoice(self, request, queryset):
-        for obj in queryset:
-            invoice = obj.create_invoice()
-            self.message_user(request, _("Invoice created"))
-            response = HttpResponseRedirect('/admin/crm/invoice/' + str(invoice.id))
-        return response
-
-    create_invoice.short_description = _("Create Invoice")
-
-    def save_model(self, request, obj, form, change):
-        if (change == True):
-            obj.last_modified_by = request.user
-        else:
-            obj.last_modified_by = request.user
-            obj.staff = request.user
-        obj.save()
-
-    actions = ['create_quote', 'create_invoice', 'create_purchase_order']
-    pluginProcessor = PluginProcessor()
-    actions.extend(pluginProcessor.getPluginAdditions("contractActions"))
-
-
-class PurchaseOrderInlinePosition(admin.TabularInline):
-    model = PurchaseOrderPosition
-    extra = 1
-    classes = ['collapse']
-    fieldsets = (
-        ('', {
-            'fields': (
-            'position_number', 'quantity', 'unit', 'product', 'description', 'overwrite_product_price',
-            'position_price_per_unit', 'sent_on')
-        }),
-    )
-    allow_add = True
-
-
-class InlineBookings(admin.TabularInline):
-    model = Booking
-    extra = 1
-    classes = ['collapse']
-    fieldsets = (
-        ('Basics', {
-            'fields': ('fromAccount', 'toAccount', 'description', 'amount', 'bookingDate', 'staff', 'bookingReference',)
-        }),
-    )
-    allow_add = False
-
-
-class OptionInvoice(admin.ModelAdmin):
-    list_display = (
-    'id', 'description', 'contract', 'customer', 'payable_until', 'status', 'currency', 'staff',
-    'last_calculated_price', 'last_calculated_tax', 'last_pricing_date', 'last_modification', 'last_modified_by', 'last_print_date')
-    list_display_links = ('id', )
-    list_filter = ('customer', 'contract', 'staff', 'status', 'currency', 'last_modification')
-    ordering = ('id',)
-    search_fields = ('contract__id', 'customer__name', 'currency__description')
-    fieldsets = (
-        (_('Basics'), {
-            'fields': ('contract', 'description', 'customer', 'currency', 'discount',  'payable_until', 'status', 'external_reference', 'template_set' )
-        }),
-    )
-    save_as = True
-    inlines = [SalesContractInlinePosition, SalesContractTextParagraph, SalesContractPostalAddress, SalesContractPhoneAddress,
-               SalesContractEmailAddress, InlineBookings]
-    pluginProcessor = PluginProcessor()
-    inlines.extend(pluginProcessor.getPluginAdditions("invoiceInlines"))
-
-    class PaymentForm(forms.Form):
-        paymentAmount = forms.DecimalField()
-        _selected_action = forms.CharField(widget=forms.MultipleHiddenInput)
-        paymentAccount = forms.ModelChoiceField(Account.objects.filter(accountType="A"))
-
-    def response_add(self, request, obj, post_url_continue=None):
-        new_obj = self.after_saving_model_and_related_inlines(request, obj)
-        return super(OptionInvoice, self).response_add(request, new_obj)
-
-    def response_change(self, request, obj):
-        new_obj = self.after_saving_model_and_related_inlines(request, obj)
-        return super(OptionInvoice, self).response_add(request, new_obj)
-
-    def after_saving_model_and_related_inlines(self, request, obj):
-        try:
-            Calculations.calculate_document_price(obj, date.today())
-            self.message_user(request, "Successfully calculated Prices")
-        except Product.NoPriceFound as e:
-            self.message_user(request, "Unsuccessfull in updating the Prices " + e.__str__(), level=messages.ERROR)
-        return obj
-
-    def save_model(self, request, obj, form, change):
-        if (change == True):
-            obj.last_modified_by = request.user
-        else:
-            obj.last_modified_by = request.user
-            obj.staff = request.user
-        obj.save()
-
-    def recalculatePrices(self, request, queryset):
-        try:
-            for obj in queryset:
-                obj.recalculate_prices(date.today())
-            self.message_user(request, "Successfully recalculated Prices")
-        except Product.NoPriceFound as e:
-            self.message_user(request, "Unsuccessfull in updating the Prices " + e.__str__(), level=messages.ERROR)
-            return;
-
-    recalculatePrices.short_description = _("Recalculate Prices")
-
-    def create_pdf(self, request, queryset):
-        for obj in queryset:
-            response = export_pdf(self, request, obj, "/admin/crm/invoice/")
-            return response
-
-    create_pdf.short_description = _("Create PDF of Invoice")
-
-    def registerInvoiceInAccounting(self, request, queryset):
-        try:
-            for obj in queryset:
-                obj.registerinvoiceinaccounting(request)
-            self.message_user(request, _("Successfully registered Invoice in the Accounting"))
-            return;
-        except OpenInterestAccountMissing as e:
-            self.message_user(request, "Did not register Invoice in Accounting: " + e.__str__(), level=messages.ERROR)
-            return;
-        except IncompleteInvoice as e:
-            self.message_user(request, "Did not register Invoice in Accounting: " + e.__str__(), level=messages.ERROR)
-            return;
-
-    registerInvoiceInAccounting.short_description = _("Register Invoice in Accounting")
-
-    # def unregisterInvoiceInAccounting(self, request, queryset):
-    # for obj in queryset:
-    # obj.createPDF(deliveryorder=True)
-    # self.message_user(request, _("Successfully unregistered Invoice in the Accounting"))
-    # unregisterInvoiceInAccounting.short_description = _("Unregister Invoice in Accounting")
-
-    def registerPaymentInAccounting(self, request, queryset):
-        form = None
-        if request.POST.get('post'):
-            if 'cancel' in request.POST:
-                self.message_user(request, _("Canceled registeration of payment in the accounting"), level=messages.ERROR)
-                return
-            elif 'register' in request.POST:
-                form = self.PaymentForm(request.POST)
-                if form.is_valid():
-                    paymentAmount = form.cleaned_data['paymentAmount']
-                    paymentAccount = form.cleaned_data['paymentAccount']
-                    for obj in queryset:
-                        obj.registerpaymentinaccounting(request, paymentAmount, paymentAccount)
-                    self.message_user(request, _("Successfully registered Payment in the Accounting"))
-                    return HttpResponseRedirect(request.get_full_path())
-        else:
-            form = self.PaymentForm
-            c = {'action_checkbox_name': helpers.ACTION_CHECKBOX_NAME, 'queryset': queryset, 'form': form}
-            c.update(csrf(request))
-            return render(request, 'crm/admin/registerPayment.html', c)
-
-
-    registerPaymentInAccounting.short_description = _("Register Payment in Accounting")
-
-    actions = ['recalculatePrices', 'create_pdf', 'registerInvoiceInAccounting',
-               'unregisterInvoiceInAccounting', 'registerPaymentInAccounting',]
-    pluginProcessor = PluginProcessor()
-    actions.extend(pluginProcessor.getPluginAdditions("invoiceActions"))
-
-
-class OptionQuote(admin.ModelAdmin):
-    list_display = (
-    'id', 'description', 'contract', 'customer', 'currency', 'valid_until', 'status', 'staff', 'last_modified_by',
-    'last_calculated_price', 'last_calculated_tax', 'last_pricing_date', 'last_modification', 'last_print_date')
-    list_display_links = ('id',)
-    list_filter = ('customer', 'contract', 'currency', 'staff', 'status', 'last_modification')
-    ordering = ('id',)
-    search_fields = ('contract__id', 'customer__name', 'currency__description')
-
-    fieldsets = (
-        (_('Basics'), {
-            'fields': ('contract', 'description', 'customer', 'currency', 'discount', 'valid_until', 'staff', 'status', 'external_reference')
-        }),
-    )
-    save_as = True
-    inlines = [SalesContractInlinePosition, SalesContractTextParagraph, SalesContractPostalAddress, SalesContractPhoneAddress,
-               SalesContractEmailAddress]
-    pluginProcessor = PluginProcessor()
-    inlines.extend(pluginProcessor.getPluginAdditions("quoteInlines"))
-
-    def response_add(self, request, new_object):
-        obj = self.after_saving_model_and_related_inlines(request, new_object)
-        return super(OptionQuote, self).response_add(request, obj)
-
-    def response_change(self, request, new_object):
-        obj = self.after_saving_model_and_related_inlines(request, new_object)
-        return super(OptionQuote, self).response_change(request, obj)
-
-    def after_saving_model_and_related_inlines(self, request, obj):
-        try:
-            Calculations.calculate_document_price(obj, date.today())
-            self.message_user(request, "Successfully calculated Prices")
-        except Product.NoPriceFound as e:
-            self.message_user(request, "Unsuccessful in updating the Prices " + e.__str__(), level=messages.ERROR)
-        return obj
-
-    def save_model(self, request, obj, form, change):
-        if (change == True):
-            obj.last_modified_by = request.user
-        else:
-            obj.last_modified_by = request.user
-            obj.staff = request.user
-        obj.save()
-
-    def recalculate_prices(self, request, queryset):
-        for obj in queryset:
-            self.after_saving_model_and_related_inlines(request, obj)
-        return;
-
-    recalculate_prices.short_description = _("Recalculate Prices")
-
-    def create_invoice(self, request, queryset):
-        for obj in queryset:
-            invoice = obj.create_invoice()
-            self.message_user(request, _("Invoice created"))
-            response = HttpResponseRedirect('/admin/crm/invoice/' + str(invoice.id))
-        return response
-
-    create_invoice.short_description = _("Create Invoice")
-
-    def create_pdf(self, request, queryset):
-        for obj in queryset:
-            response = export_pdf(self, request, obj, "/admin/crm/quote/")
-            return response
-
-    create_pdf.short_description = _("Create PDF of Quote")
-
-    actions = ['recalculate_prices', 'create_invoice', 'create_pdf']
-    pluginProcessor = PluginProcessor()
-    inlines.extend(pluginProcessor.getPluginAdditions("quoteActions"))
-
-
-class OptionPurchaseOrder(admin.ModelAdmin):
-    list_display = (
-    'id', 'description', 'contract', 'supplier', 'status', 'currency', 'staff', 'last_modified_by',
-    'last_calculated_price', 'last_calculated_tax', 'last_pricing_date', 'last_modification', 'last_print_date')
-    list_display_links = ('id',)
-    list_filter = ('supplier', 'contract', 'staff', 'status', 'currency', 'last_modification')
-    ordering = ('id',)
-    search_fields = ('contract__id', 'supplier__name', 'currency_description')
-
-    fieldsets = (
-        (_('Basics'), {
-            'fields': ('contract', 'description', 'supplier', 'currency', 'status', 'external_reference')
-        }),
-    )
-
-    def response_add(self, request, new_object):
-        obj = self.after_saving_model_and_related_inlines(request, new_object)
-        return super(OptionPurchaseOrder, self).response_add(request, obj)
-
-    def response_change(self, request, new_object):
-        obj = self.after_saving_model_and_related_inlines(request, new_object)
-        return super(OptionPurchaseOrder, self).response_change(request, obj)
-
-    def after_saving_model_and_related_inlines(self, request, obj):
-        try:
-            Calculations.calculate_document_price(obj, date.today())
-            self.message_user(request, "Successfully calculated Prices")
-        except Product.NoPriceFound as e:
-            self.message_user(request, "Unsuccessful in updating the Prices " + e.__str__(), level=messages.ERROR)
-        return obj
-
-    def save_model(self, request, obj, form, change):
-        if (change == True):
-            obj.last_modified_by = request.user
-        else:
-            obj.last_modified_by = request.user
-            obj.staff = request.user
-        obj.save()
-
-    def recalculatePrices(self, request, queryset):
-        for obj in queryset:
-            self.after_saving_model_and_related_inlines(request, obj)
-        return;
-
-    recalculatePrices.short_description = _("Recalculate Prices")
-
-    def createPurchseOrderPDF(self, request, queryset):
-        for obj in queryset:
-            response = export_pdf(self, request, obj, "/admin/crm/purchaseorder/")
-            return response
-
-    createPurchseOrderPDF.short_description = _("Create PDF of Purchase Order")
-
-    actions = ['createPurchseOrderPDF']
-    pluginProcessor = PluginProcessor()
-    actions.extend(pluginProcessor.getPluginAdditions("purchaseOrderActions"))
-
-    save_as = True
-    inlines = [PurchaseOrderInlinePosition, PurchaseOrderPostalAddress, PurchaseOrderPhoneAddress,
-               PurchaseOrderEmailAddress]
-    pluginProcessor = PluginProcessor()
-    inlines.extend(pluginProcessor.getPluginAdditions("purchaseOrderInlines"))
-
-
-class ProductPrice(admin.TabularInline):
-    model = Price
-    extra = 1
-    classes = ['collapse']
-    fieldsets = (
-        ('', {
-            'fields': ('price', 'currency', 'unit', 'valid_from', 'valid_until', 'customer_group')
-        }),
-    )
-    allow_add = True
-
-
-class ProductUnitTransform(admin.TabularInline):
-    model = UnitTransform
-    extra = 1
-    classes = ['collapse']
-    fieldsets = (
-        ('', {
-            'fields': ('from_unit', 'to_unit', 'factor',)
-        }),
-    )
-    allow_add = True
-
-
-class OptionProduct(admin.ModelAdmin):
-    list_display = ('product_number', 'title', 'default_unit', 'tax', 'accounting_product_categorie')
-    list_display_links = ('product_number',)
-    fieldsets = (
-        (_('Basics'), {
-            'fields': ('product_number', 'title', 'description', 'default_unit', 'tax', 'accounting_product_categorie')
-        }),)
-    inlines = [ProductPrice, ProductUnitTransform]
-
-
-class ContactPostalAddress(admin.StackedInline):
-    model = PostalAddressForContact
-    extra = 1
-    classes = ['collapse']
-    fieldsets = (
-        ('Basics', {
-            'fields': (
-            'prefix', 'prename', 'name', 'addressline1', 'addressline2', 'addressline3', 'addressline4', 'zipcode',
-            'town', 'state', 'country', 'purpose')
-        }),
-    )
-    allow_add = True
-
-
-class ContactPhoneAddress(admin.TabularInline):
-    model = PhoneAddressForContact
-    extra = 1
-    classes = ['collapse']
-    fieldsets = (
-        ('Basics', {
-            'fields': ('phone', 'purpose',)
-        }),
-    )
-    allow_add = True
-
-
-class ContactEmailAddress(admin.TabularInline):
-    model = EmailAddressForContact
-    extra = 1
-    classes = ['collapse']
-    fieldsets = (
-        ('Basics', {
-            'fields': ('email', 'purpose',)
-        }),
-    )
-    allow_add = True
-
-
-class OptionCustomer(admin.ModelAdmin):
-    list_display = ('id', 'name', 'defaultCustomerBillingCycle',)
-    fieldsets = (('', {'fields': ('name', 'defaultCustomerBillingCycle', 'ismemberof',)}),)
-    allow_add = True
-    ordering = ('id',)
-    search_fields = ('id', 'name')
-    inlines = [ContactPostalAddress, ContactPhoneAddress, ContactEmailAddress]
-    pluginProcessor = PluginProcessor()
-    inlines.extend(pluginProcessor.getPluginAdditions("customerInline"))
-
-    def createContract(self, request, queryset):
-        for obj in queryset:
-            contract = obj.createContract(request)
-            response = HttpResponseRedirect('/admin/crm/contract/' + str(contract.id))
-            return response
-
-    createContract.short_description = _("Create Contract")
-
-    @staticmethod
-    def createQuote(self, request, queryset):
-        for obj in queryset:
-            quote = obj.createQuote()
-            response = HttpResponseRedirect('/admin/crm/quote/' + str(quote.id))
-        return response
-
-    createQuote.short_description = _("Create Quote")
-
-    @staticmethod
-    def createInvoice(self, request, queryset):
-        for obj in queryset:
-            invoice = obj.createInvoice()
-            response = HttpResponseRedirect('/admin/crm/invoice/' + str(invoice.id))
-        return response
-
-    createInvoice.short_description = _("Create Invoice")
-
-    def save_model(self, request, obj, form, change):
-        if (change == True):
-            obj.lastmodifiedby = request.user
-        else:
-            obj.lastmodifiedby = request.user
-            obj.staff = request.user
-        obj.save()
-
-    actions = ['createContract', 'createInvoice', 'createQuote']
-    pluginProcessor = PluginProcessor()
-    inlines.extend(pluginProcessor.getPluginAdditions("customerActions"))
-
-
-class OptionCustomerGroup(admin.ModelAdmin):
-    list_display = ('id', 'name')
-    fieldsets = (('', {'fields': ('name',)}),)
-    allow_add = True
-
-
-class OptionSupplier(admin.ModelAdmin):
-    list_display = ('id', 'name', 'offersShipmentToCustomers')
-    fieldsets = (('', {'fields': ('name', 'offersShipmentToCustomers')}),)
-    inlines = [ContactPostalAddress, ContactPhoneAddress, ContactEmailAddress]
-    allow_add = True
-
-    def save_model(self, request, obj, form, change):
-        if (change == True):
-            obj.lastmodifiedby = request.user
-        else:
-            obj.lastmodifiedby = request.user
-            obj.staff = request.user
-        obj.save()
-
-
-class OptionUnit(admin.ModelAdmin):
-    list_display = ('id', 'description', 'short_name', 'is_a_fraction_of', 'fraction_factor_to_next_higher_unit')
-    fieldsets = (('', {'fields': ('description', 'short_name', 'is_a_fraction_of', 'fraction_factor_to_next_higher_unit')}),)
-    allow_add = True
-
-
-class OptionCurrency(admin.ModelAdmin):
-    list_display = ('id', 'description', 'short_name', 'rounding')
-    fieldsets = (('', {'fields': ('description', 'short_name', 'rounding')}),)
-    allow_add = True
-
-
-class OptionTax(admin.ModelAdmin):
-    list_display = ('id', 'tax_rate', 'name', 'account_activa', 'account_passiva')
-    fieldsets = (('', {'fields': ('tax_rate', 'name', 'account_activa', 'account_passiva')}),)
-    allow_add = True
-
-
-class OptionCustomerBillingCycle(admin.ModelAdmin):
-    list_display = ('id', 'timeToPaymentDate', 'name')
-    fieldsets = (('', {'fields': ('timeToPaymentDate', 'name',)}),)
-    allow_add = True
+from koalixcrm.crm.documents.quote import OptionQuote
+from koalixcrm.crm.documents.invoice import OptionInvoice
+from koalixcrm.crm.documents.purchaseorder import OptionPurchaseOrder
+from koalixcrm.crm.documents.contract import OptionContract
+from koalixcrm.crm.product.tax import OptionTax
+from koalixcrm.crm.product.unit import OptionUnit
+from koalixcrm.crm.product.product import OptionProduct
+from koalixcrm.crm.product.currency import OptionCurrency
+from koalixcrm.crm.contact.customer import OptionCustomer
+from koalixcrm.crm.contact.supplier import OptionSupplier
+from koalixcrm.crm.contact.customergroup import OptionCustomerGroup
+from koalixcrm.crm.contact.customerbillingcycle import OptionCustomerBillingCycle
 
 
 admin.site.register(Customer, OptionCustomer)
