+++ koalixcrm/crm/contact/customer.py
@@ -1,9 +1,14 @@
 # -*- coding: utf-8 -*-
 
 from django.db import models
+from django.contrib import admin
 from django.utils.translation import ugettext as _
+from koalixcrm.plugin import *
 from koalixcrm import djangoUserExtension
 from koalixcrm.crm.contact.contact import Contact
+from koalixcrm.crm.contact.contact import ContactPostalAddress
+from koalixcrm.crm.contact.contact import ContactPhoneAddress
+from koalixcrm.crm.contact.contact import ContactEmailAddress
 
 import koalixcrm.crm.documents.contract
 
@@ -44,3 +49,52 @@ class Customer(Contact):
 
     def __str__(self):
         return str(self.id) + ' ' + self.name
+
+
+class OptionCustomer(admin.ModelAdmin):
+    list_display = ('id', 'name', 'defaultCustomerBillingCycle',)
+    fieldsets = (('', {'fields': ('name', 'defaultCustomerBillingCycle', 'ismemberof',)}),)
+    allow_add = True
+    ordering = ('id',)
+    search_fields = ('id', 'name')
+    inlines = [ContactPostalAddress, ContactPhoneAddress, ContactEmailAddress]
+    pluginProcessor = PluginProcessor()
+    inlines.extend(pluginProcessor.getPluginAdditions("customerInline"))
+
+    def createContract(self, request, queryset):
+        for obj in queryset:
+            contract = obj.createContract(request)
+            response = HttpResponseRedirect('/admin/crm/contract/' + str(contract.id))
+            return response
+
+    createContract.short_description = _("Create Contract")
+
+    @staticmethod
+    def createQuote(self, request, queryset):
+        for obj in queryset:
+            quote = obj.createQuote()
+            response = HttpResponseRedirect('/admin/crm/quote/' + str(quote.id))
+        return response
+
+    createQuote.short_description = _("Create Quote")
+
+    @staticmethod
+    def createInvoice(self, request, queryset):
+        for obj in queryset:
+            invoice = obj.createInvoice()
+            response = HttpResponseRedirect('/admin/crm/invoice/' + str(invoice.id))
+        return response
+
+    createInvoice.short_description = _("Create Invoice")
+
+    def save_model(self, request, obj, form, change):
+        if (change == True):
+            obj.lastmodifiedby = request.user
+        else:
+            obj.lastmodifiedby = request.user
+            obj.staff = request.user
+        obj.save()
+
+    actions = ['createContract', 'createInvoice', 'createQuote']
+    pluginProcessor = PluginProcessor()
+    inlines.extend(pluginProcessor.getPluginAdditions("customerActions"))
