+++ koalixcrm/crm/contact/customergroup.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from django.db import models
+from django.contrib import admin
 from django.utils.translation import ugettext as _
 
 
@@ -14,3 +15,9 @@ class CustomerGroup(models.Model):
         app_label = "crm"
         verbose_name = _('Customer Group')
         verbose_name_plural = _('Customer Groups')
+
+
+class OptionCustomerGroup(admin.ModelAdmin):
+    list_display = ('id', 'name')
+    fieldsets = (('', {'fields': ('name',)}),)
+    allow_add = True
