+++ koalixcrm/crm/documents/contract.py
@@ -2,7 +2,10 @@
 
 from datetime import *
 from django.db import models
+from django.contrib import admin
 from django.utils.translation import ugettext as _
+
+from koalixcrm.plugin import *
 from koalixcrm.crm.contact.phoneaddress import PhoneAddress
 from koalixcrm.crm.contact.emailaddress import EmailAddress
 from koalixcrm.crm.contact.postaladdress import PostalAddress
@@ -11,6 +14,86 @@ from koalixcrm.crm.documents.quote import Quote
 from koalixcrm.crm.documents.purchaseorder import PurchaseOrder
 from koalixcrm.globalSupportFunctions import xstr
 from koalixcrm.crm.const.purpose import *
+from koalixcrm.crm.documents.invoice import InlineInvoice
+from koalixcrm.crm.documents.quote import InlineQuote
+from koalixcrm.crm.documents.purchaseorder import InlinePurchaseOrder
+
+
+class PostalAddressForContract(PostalAddress):
+    purpose = models.CharField(verbose_name=_("Purpose"), max_length=1, choices=PURPOSESADDRESSINCONTRACT)
+    contract = models.ForeignKey('Contract')
+
+    class Meta:
+        app_label = "crm"
+        verbose_name = _('Postal Address For Contracts')
+        verbose_name_plural = _('Postal Address For Contracts')
+
+    def __str__(self):
+        return xstr(self.prename) + ' ' + xstr(self.name) + ' ' + xstr(self.addressline1)
+
+
+class PhoneAddressForContract(PhoneAddress):
+    purpose = models.CharField(verbose_name=_("Purpose"), max_length=1, choices=PURPOSESADDRESSINCONTRACT)
+    contract = models.ForeignKey('Contract')
+
+    class Meta:
+        app_label = "crm"
+        verbose_name = _('Phone Address For Contracts')
+        verbose_name_plural = _('Phone Address For Contracts')
+
+    def __str__(self):
+        return str(self.phone)
+
+
+class EmailAddressForContract(EmailAddress):
+    purpose = models.CharField(verbose_name=_("Purpose"), max_length=1, choices=PURPOSESADDRESSINCONTRACT)
+    contract = models.ForeignKey('Contract')
+
+    class Meta:
+        app_label = "crm"
+        verbose_name = _('Email Address For Contracts')
+        verbose_name_plural = _('Email Address For Contracts')
+
+    def __str__(self):
+        return str(self.email)
+
+
+class ContractPostalAddress(admin.StackedInline):
+    model = PostalAddressForContract
+    extra = 1
+    classes = ['collapse']
+    fieldsets = (
+        ('Basics', {
+            'fields': (
+            'prefix', 'prename', 'name', 'addressline1', 'addressline2', 'addressline3', 'addressline4', 'zipcode',
+            'town', 'state', 'country', 'purpose'),
+        }),
+    )
+    allow_add = True
+
+
+class ContractPhoneAddress(admin.TabularInline):
+    model = PhoneAddressForContract
+    extra = 1
+    classes = ['collapse']
+    fieldsets = (
+        ('Basics', {
+            'fields': ('phone', 'purpose',)
+        }),
+    )
+    allow_add = True
+
+
+class ContractEmailAddress(admin.TabularInline):
+    model = EmailAddressForContract
+    extra = 1
+    classes = ['collapse']
+    fieldsets = (
+        ('Basics', {
+            'fields': ('email', 'purpose',)
+        }),
+    )
+    allow_add = True
 
 
 class Contract(models.Model):
@@ -58,40 +141,58 @@ class Contract(models.Model):
         return _("Contract") + " " + str(self.id)
 
 
-class PostalAddressForContract(PostalAddress):
-    purpose = models.CharField(verbose_name=_("Purpose"), max_length=1, choices=PURPOSESADDRESSINCONTRACT)
-    contract = models.ForeignKey(Contract)
+class OptionContract(admin.ModelAdmin):
+    list_display = ('id', 'description', 'default_customer', 'default_supplier', 'staff', 'default_currency')
+    list_display_links = ('id',)
+    list_filter = ('default_customer', 'default_supplier', 'staff', 'default_currency')
+    ordering = ('id', )
+    search_fields = ('id', 'contract')
+    fieldsets = (
+        (_('Basics'), {
+            'fields': ('description', 'default_customer', 'staff', 'default_supplier', 'default_currency', 'default_template_set')
+        }),
+    )
+    inlines = [ContractPostalAddress, ContractPhoneAddress, ContractEmailAddress, InlineQuote, InlineInvoice,
+               InlinePurchaseOrder]
+    pluginProcessor = PluginProcessor()
+    inlines.extend(pluginProcessor.getPluginAdditions("contractInlines"))
 
-    class Meta:
-        app_label = "crm"
-        verbose_name = _('Postal Address For Contracts')
-        verbose_name_plural = _('Postal Address For Contracts')
+    def create_purchase_order(self, request, queryset):
+        for obj in queryset:
+            purchase_order = obj.create_purchase_order()
+            self.message_user(request, _("PurchaseOrder created"))
+            response = HttpResponseRedirect('/admin/crm/purchaseorder/' + str(purchase_order.id))
+        return response
 
-    def __str__(self):
-        return xstr(self.prename) + ' ' + xstr(self.name) + ' ' + xstr(self.addressline1)
+        create_purchase_order.short_description = _("Create Purchaseorder")
 
+    def create_quote(self, request, queryset):
+        for obj in queryset:
+            quote = obj.create_quote()
+            self.message_user(request, _("Quote created"))
+            response = HttpResponseRedirect('/admin/crm/quote/' + str(quote.id))
+        return response
 
-class PhoneAddressForContract(PhoneAddress):
-    purpose = models.CharField(verbose_name=_("Purpose"), max_length=1, choices=PURPOSESADDRESSINCONTRACT)
-    contract = models.ForeignKey(Contract)
+    create_quote.short_description = _("Create Quote")
 
-    class Meta:
-        app_label = "crm"
-        verbose_name = _('Phone Address For Contracts')
-        verbose_name_plural = _('Phone Address For Contracts')
+    def create_invoice(self, request, queryset):
+        for obj in queryset:
+            invoice = obj.create_invoice()
+            self.message_user(request, _("Invoice created"))
+            response = HttpResponseRedirect('/admin/crm/invoice/' + str(invoice.id))
+        return response
 
-    def __str__(self):
-        return str(self.phone)
+    create_invoice.short_description = _("Create Invoice")
 
+    def save_model(self, request, obj, form, change):
+        if (change == True):
+            obj.last_modified_by = request.user
+        else:
+            obj.last_modified_by = request.user
+            obj.staff = request.user
+        obj.save()
 
-class EmailAddressForContract(EmailAddress):
-    purpose = models.CharField(verbose_name=_("Purpose"), max_length=1, choices=PURPOSESADDRESSINCONTRACT)
-    contract = models.ForeignKey(Contract)
+    actions = ['create_quote', 'create_invoice', 'create_purchase_order']
+    pluginProcessor = PluginProcessor()
+    actions.extend(pluginProcessor.getPluginAdditions("contractActions"))
 
-    class Meta:
-        app_label = "crm"
-        verbose_name = _('Email Address For Contracts')
-        verbose_name_plural = _('Email Address For Contracts')
-
-    def __str__(self):
-        return str(self.email)
