+++ koalixcrm/crm/documents/deliverynote.py
@@ -9,13 +9,14 @@ import koalixcrm.crm.documents.pdfexport
 
 
 class DeliveryNote(SalesContract):
+    derived_from_quote = models.ForeignKey("Quote", blank=True, null=True)
     derived_from_invoice = models.ForeignKey("Invoice", blank=True, null=True)
     status = models.CharField(max_length=1, choices=DELIVERYNOTESTATUS)
 
-    def createPDF(self):
+    def create_pdf(self):
         self.last_print_date = datetime.now()
         self.save()
-        return koalixcrm.crm.documents.pdfexport.PDFExport.createPDF(self)
+        return koalixcrm.crm.documents.pdfexport.PDFExport.create_pdf(self)
 
     def __str__(self):
         return _("Delivery Note") + ": " + str(self.id) + " " + _("from Contract") + ": " + str(self.contract.id)
