+++ koalixcrm/crm/documents/invoice.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
 
 from datetime import *
+from django import forms
 from django.db import models
+from django.contrib import admin
 from django.utils.translation import ugettext as _
 from koalixcrm.crm.const.status import *
 from koalixcrm.crm.exceptions import *
@@ -11,6 +13,14 @@ from koalixcrm.crm.documents.salescontract import TextParagraphInSalesContract
 from koalixcrm.crm.documents.salescontractposition import SalesContractPosition
 from koalixcrm.djangoUserExtension.models import TextParagraphInDocumentTemplate
 import koalixcrm.crm.documents.pdfexport
+from koalixcrm.plugin import *
+from koalixcrm.crm.documents.salescontract import SalesContractTextParagraph
+from koalixcrm.crm.documents.salescontract import SalesContractPostalAddress
+from koalixcrm.crm.documents.salescontract import SalesContractPhoneAddress
+from koalixcrm.crm.documents.salescontract import SalesContractEmailAddress
+from koalixcrm.crm.documents.salescontractposition import SalesContractInlinePosition
+from koalixcrm.accounting.admin import InlineBookings
+from koalixcrm.accounting.models import Account
 
 
 class Invoice(SalesContract):
@@ -32,15 +42,11 @@ class Invoice(SalesContract):
             self.customer = calling_model.default_customer
             self.currency = calling_model.default_currency
             self.description = calling_model.description
-            self.template_set = calling_model.default_template_set
+            self.template_set = calling_model.default_template_set.invoice_template
             self.discount = 0
         elif type(calling_model) == koalixcrm.crm.documents.quote.Quote:
-            self.contract = calling_model.contract
             self.derived_from_quote = calling_model
-            self.customer = calling_model.customer
-            self.currency = calling_model.currency
-            self.discount = calling_model.discount
-            self.description = calling_model.description
+            self.copy_sales_contract(calling_model)
 
         self.status = 'C'
         self.payable_until = date.today() + \
@@ -64,12 +70,6 @@ class Invoice(SalesContract):
             for quote_paragraph in list(quote_paragraphs):
                 invoice_paragraph = TextParagraphInSalesContract()
                 invoice_paragraph.create_paragraph(quote_paragraph, self)
-
-    def get_fop_config_file(self):
-        return self.template_set.invoice_template.fop_config_file
-
-    def get_xsl_file(self):
-        return self.template_set.invoice_template.xsl_file
 
     def is_complete_with_price(self):
         """ Checks whether the Invoice is completed with a price, in case the invoice
@@ -124,10 +124,10 @@ class Invoice(SalesContract):
         booking.lastmodifiedby = request.user
         booking.save()
 
-    def createPDF(self):
+    def create_pdf(self):
         self.last_print_date = datetime.now()
         self.save()
-        return koalixcrm.crm.documents.pdfexport.PDFExport.createPDF(self)
+        return koalixcrm.crm.documents.pdfexport.PDFExport.create_pdf(self)
 
     def __str__(self):
         return _("Invoice") + ": " + str(self.id) + " " + _("from Contract") + ": " + str(self.contract.id)
@@ -136,3 +136,142 @@ class Invoice(SalesContract):
         app_label = "crm"
         verbose_name = _('Invoice')
         verbose_name_plural = _('Invoices')
+
+
+class OptionInvoice(admin.ModelAdmin):
+    list_display = (
+    'id', 'description', 'contract', 'customer', 'payable_until', 'status', 'currency', 'staff',
+    'last_calculated_price', 'last_calculated_tax', 'last_pricing_date', 'last_modification', 'last_modified_by', 'last_print_date')
+    list_display_links = ('id', )
+    list_filter = ('customer', 'contract', 'staff', 'status', 'currency', 'last_modification')
+    ordering = ('id',)
+    search_fields = ('contract__id', 'customer__name', 'currency__description')
+    fieldsets = (
+        (_('Basics'), {
+            'fields': ('contract', 'description', 'customer', 'currency', 'discount',  'payable_until', 'status', 'external_reference', 'template_set' )
+        }),
+    )
+    save_as = True
+    inlines = [SalesContractInlinePosition, SalesContractTextParagraph, SalesContractPostalAddress, SalesContractPhoneAddress,
+               SalesContractEmailAddress, InlineBookings]
+    pluginProcessor = PluginProcessor()
+    inlines.extend(pluginProcessor.getPluginAdditions("invoiceInlines"))
+
+    class PaymentForm(forms.Form):
+        paymentAmount = forms.DecimalField()
+        _selected_action = forms.CharField(widget=forms.MultipleHiddenInput)
+        paymentAccount = forms.ModelChoiceField(Account.objects.filter(accountType="A"))
+
+    def response_add(self, request, obj, post_url_continue=None):
+        new_obj = self.after_saving_model_and_related_inlines(request, obj)
+        return super(OptionInvoice, self).response_add(request, new_obj)
+
+    def response_change(self, request, obj):
+        new_obj = self.after_saving_model_and_related_inlines(request, obj)
+        return super(OptionInvoice, self).response_add(request, new_obj)
+
+    def after_saving_model_and_related_inlines(self, request, obj):
+        try:
+            Calculations.calculate_document_price(obj, date.today())
+            self.message_user(request, "Successfully calculated Prices")
+        except Product.NoPriceFound as e:
+            self.message_user(request, "Unsuccessfull in updating the Prices " + e.__str__(), level=messages.ERROR)
+        return obj
+
+    def save_model(self, request, obj, form, change):
+        if (change == True):
+            obj.last_modified_by = request.user
+        else:
+            obj.last_modified_by = request.user
+            obj.staff = request.user
+        obj.save()
+
+    def recalculatePrices(self, request, queryset):
+        try:
+            for obj in queryset:
+                obj.recalculate_prices(date.today())
+            self.message_user(request, "Successfully recalculated Prices")
+        except Product.NoPriceFound as e:
+            self.message_user(request, "Unsuccessfull in updating the Prices " + e.__str__(), level=messages.ERROR)
+            return;
+
+    recalculatePrices.short_description = _("Recalculate Prices")
+
+    def create_pdf(self, request, queryset):
+        for obj in queryset:
+            response = export_pdf(self, request, obj, "/admin/crm/invoice/")
+            return response
+
+    create_pdf.short_description = _("Create PDF of Invoice")
+
+    def registerInvoiceInAccounting(self, request, queryset):
+        try:
+            for obj in queryset:
+                obj.registerinvoiceinaccounting(request)
+            self.message_user(request, _("Successfully registered Invoice in the Accounting"))
+            return;
+        except OpenInterestAccountMissing as e:
+            self.message_user(request, "Did not register Invoice in Accounting: " + e.__str__(), level=messages.ERROR)
+            return;
+        except IncompleteInvoice as e:
+            self.message_user(request, "Did not register Invoice in Accounting: " + e.__str__(), level=messages.ERROR)
+            return;
+
+    registerInvoiceInAccounting.short_description = _("Register Invoice in Accounting")
+
+    # def unregisterInvoiceInAccounting(self, request, queryset):
+    # for obj in queryset:
+    # obj.createPDF(deliveryorder=True)
+    # self.message_user(request, _("Successfully unregistered Invoice in the Accounting"))
+    # unregisterInvoiceInAccounting.short_description = _("Unregister Invoice in Accounting")
+
+    def registerPaymentInAccounting(self, request, queryset):
+        form = None
+        if request.POST.get('post'):
+            if 'cancel' in request.POST:
+                self.message_user(request, _("Canceled registeration of payment in the accounting"), level=messages.ERROR)
+                return
+            elif 'register' in request.POST:
+                form = self.PaymentForm(request.POST)
+                if form.is_valid():
+                    paymentAmount = form.cleaned_data['paymentAmount']
+                    paymentAccount = form.cleaned_data['paymentAccount']
+                    for obj in queryset:
+                        obj.registerpaymentinaccounting(request, paymentAmount, paymentAccount)
+                    self.message_user(request, _("Successfully registered Payment in the Accounting"))
+                    return HttpResponseRedirect(request.get_full_path())
+        else:
+            form = self.PaymentForm
+            c = {'action_checkbox_name': helpers.ACTION_CHECKBOX_NAME, 'queryset': queryset, 'form': form}
+            c.update(csrf(request))
+            return render(request, 'crm/admin/registerPayment.html', c)
+
+
+    registerPaymentInAccounting.short_description = _("Register Payment in Accounting")
+
+    actions = ['recalculatePrices', 'create_pdf', 'registerInvoiceInAccounting',
+               'unregisterInvoiceInAccounting', 'registerPaymentInAccounting',]
+    pluginProcessor = PluginProcessor()
+    actions.extend(pluginProcessor.getPluginAdditions("invoiceActions"))
+
+
+class InlineInvoice(admin.TabularInline):
+    model = Invoice
+    classes = ['collapse']
+    show_change_link = True
+    can_delete = True
+    extra = 1
+    readonly_fields = (
+    'last_pricing_date', 'last_calculated_price', 'last_calculated_tax', 'description', 'contract', 'customer',
+    'payable_until', 'status')
+    fieldsets = (
+        (_('Basics'), {
+            'fields': ('description', 'contract', 'customer', 'payable_until', 'status'),
+        }),
+        (_('Advanced (not editable)'), {
+            'classes': ('collapse',),
+            'fields': ('last_pricing_date', 'last_calculated_price', 'last_calculated_tax',)
+        }),
+    )
+
+    allow_add = False
