+++ koalixcrm/crm/documents/purchaseconfirmation.py
@@ -10,13 +10,32 @@ import koalixcrm.crm.documents.pdfexport
 
 
 class PurchaseConfirmation(SalesContract):
-    valid_until = models.DateField(verbose_name=_("Valid until"))
-    status = models.CharField(max_length=1, choices=QUOTESTATUS, verbose_name=_('Status'))
+    derived_from_quote = models.ForeignKey("Quote", blank=True, null=True)
 
-    def createPDF(self):
+    def create_purchase_confirmation(self, calling_model):
+        """Checks which model was calling the function. Depending on the calling
+        model, the function sets up a purchase confirmation. On success, the
+        purchase confirmation is saved.
+        At the moment only the koalixcrm.crm.documents.contract.Contract and
+        koalixcrm.crm.documents.quote.Quote are allowed to call this function"""
+
+        self.staff = calling_model.staff
+        if type(calling_model) == koalixcrm.crm.documents.contract.Contract:
+            self.contract = calling_model
+            self.customer = calling_model.default_customer
+            self.currency = calling_model.default_currency
+            self.description = calling_model.description
+            self.template_set = calling_model.default_template_set.purchase_confirmation_template
+            self.discount = 0
+        elif type(calling_model) == koalixcrm.crm.documents.quote.Quote:
+            self.derived_from_quote = calling_model
+            self.copy_sales_contract(calling_model)
+
+
+    def create_pdf(self):
         self.last_print_date = datetime.now()
         self.save()
-        return koalixcrm.crm.documents.pdfexport.PDFExport.createPDF(self)
+        return koalixcrm.crm.documents.pdfexport.PDFExport.create_pdf(self)
 
     def __str__(self):
         return _("Purchase Confirmation") + ": " + str(self.id) + " " + _("from Contract") + ": " + str(self.contract.id)
