+++ koalixcrm/crm/models.py
@@ -9,14 +9,18 @@ from koalixcrm.crm.contact.emailaddress import *
 from koalixcrm.crm.contact.phoneaddress import *
 from koalixcrm.crm.contact.supplier import *
 
-from koalixcrm.crm.documents.contract import *
-from koalixcrm.crm.documents.salescontractposition import *
-from koalixcrm.crm.documents.salescontract import *
-from koalixcrm.crm.documents.contract import *
-from koalixcrm.crm.documents.invoice import *
-from koalixcrm.crm.documents.purchaseconfirmation import *
-from koalixcrm.crm.documents.purchaseorder import *
-from koalixcrm.crm.documents.quote import *
+from koalixcrm.crm.documents.contract import Contract, PostalAddressForContract
+from koalixcrm.crm.documents.contract import PhoneAddressForContract, EmailAddressForContract
+from koalixcrm.crm.documents.salescontractposition import Position, SalesContractPosition
+from koalixcrm.crm.documents.salescontract import SalesContract, PostalAddressForSalesContract
+from koalixcrm.crm.documents.salescontract import EmailAddressForSalesContract, PhoneAddressForSalesContract
+from koalixcrm.crm.documents.salescontract import TextParagraphInSalesContract
+from koalixcrm.crm.documents.invoice import Invoice
+from koalixcrm.crm.documents.purchaseconfirmation import PurchaseConfirmation
+from koalixcrm.crm.documents.purchaseorder import PurchaseOrder, PurchaseOrderPosition
+from koalixcrm.crm.documents.purchaseorder import PhoneAddressForPurchaseOrder, PostalAddressForPurchaseOrder
+from koalixcrm.crm.documents.purchaseorder import EmailAddressForPurchaseOrder
+from koalixcrm.crm.documents.quote import Quote
 
 from koalixcrm.crm.product.currency import *
 from koalixcrm.crm.product.price import *
