+++ koalixcrm/crm/product/currency.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from django.db import models
+from django.contrib import admin
 from django.utils.translation import ugettext as _
 
 
@@ -16,3 +17,9 @@ class Currency(models.Model):
         app_label = "crm"
         verbose_name = _('Currency')
         verbose_name_plural = _('Currency')
+
+
+class OptionCurrency(admin.ModelAdmin):
+    list_display = ('id', 'description', 'short_name', 'rounding')
+    fieldsets = (('', {'fields': ('description', 'short_name', 'rounding')}),)
+    allow_add = True
