+++ koalixcrm/crm/product/tax.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from django.db import models
+from django.contrib import admin
 from django.utils.translation import ugettext as _
 
 
@@ -22,3 +23,9 @@ class Tax(models.Model):
         app_label = "crm"
         verbose_name = _('Tax')
         verbose_name_plural = _('Taxes')
+
+
+class OptionTax(admin.ModelAdmin):
+    list_display = ('id', 'tax_rate', 'name', 'account_activa', 'account_passiva')
+    fieldsets = (('', {'fields': ('tax_rate', 'name', 'account_activa', 'account_passiva')}),)
+    allow_add = True
