+++ koalixcrm/crm/product/unit.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from django.db import models
+from django.contrib import admin
 from django.utils.translation import ugettext as _
 
 
@@ -18,6 +19,12 @@ class Unit(models.Model):
         app_label = "crm"
         verbose_name = _('Unit')
         verbose_name_plural = _('Units')
+
+
+class OptionUnit(admin.ModelAdmin):
+    list_display = ('id', 'description', 'short_name', 'is_a_fraction_of', 'fraction_factor_to_next_higher_unit')
+    fieldsets = (('', {'fields': ('description', 'short_name', 'is_a_fraction_of', 'fraction_factor_to_next_higher_unit')}),)
+    allow_add = True
 
 
 class UnitTransform(models.Model):
@@ -39,3 +46,15 @@ class UnitTransform(models.Model):
         app_label = "crm"
         verbose_name = _('Unit Transfrom')
         verbose_name_plural = _('Unit Transfroms')
+
+
+class ProductUnitTransform(admin.TabularInline):
+    model = UnitTransform
+    extra = 1
+    classes = ['collapse']
+    fieldsets = (
+        ('', {
+            'fields': ('from_unit', 'to_unit', 'factor',)
+        }),
+    )
+    allow_add = True
